chore(assets): move image-related files from assets to public for Next.js project compatibility
